[SAFE-72] Adapt second part of the mutations to CASL

+++ schema/mutation/deleteRecord.ts
@@ -1,7 +1,9 @@
-import { GraphQLNonNull, GraphQLID } from "graphql";
+import { GraphQLNonNull, GraphQLID, GraphQLError } from "graphql";
 import { Record, Version } from "../../models";
 import { RecordType } from "../types";
 import mongoose from 'mongoose';
+import { AppAbility } from "../../security/defineAbilityFor";
+import errors from "../../const/errors";
 
 export default {
     /*  Delete a record, if user has permission to update associated form / resource.
@@ -12,9 +14,13 @@ export default {
         id: { type: new GraphQLNonNull(GraphQLID) },
     },
     async resolve(parent, args, context) {
-        const user = context.user;
-        const record = await Record.findById(args.id);
-        await Version.deleteMany({ _id: { $in: record.versions.map(x => mongoose.Types.ObjectId(x))}});
-        return Record.findByIdAndRemove(args.id);
+        const ability: AppAbility = context.user.ability;
+        if (ability.can('delete', 'Record')) {
+            const record = await Record.findById(args.id);
+            await Version.deleteMany({ _id: { $in: record.versions.map(x => mongoose.Types.ObjectId(x))}});
+            return Record.findByIdAndRemove(args.id);
+        } else {
+            throw new GraphQLError(errors.permissionNotGranted);
+        }
     },
 }
